parser => expansion + check_sysntax
