+++ Makefile
@@ -6,7 +6,7 @@
 #    By: aben-ham <aben-ham@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/02/22 09:30:37 by aben-ham          #+#    #+#              #
-#    Updated: 2022/02/28 13:14:50 by aben-ham         ###   ########.fr        #
+#    Updated: 2022/03/01 14:52:12 by aben-ham         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -45,12 +45,15 @@ FILES = \
 	src/signals.c \
 	parser/parser.c \
 	parser/expansion.c \
+	parser/sysntax_1.c \
+	parser/sysntax_2.c \
 	main.c
 
 #ls -d [dir/*] | grep "" | sed 's/$/ \\/g'
 #-Wall -Wextra -Werror
 CFLAGS = -g -I includes/
 #FLAGS = -lreadline -L ~/brew/opt/readline/lib  -I ~/brew/opt/readline/include
+FLAGS = -fsanitize=address
 DEPFLAGS = -MMD -MF $(@:.o=.d)
 
 NAME = minishell
