+++ main.c
@@ -6,7 +6,7 @@
 /*   By: aben-ham <aben-ham@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/22 10:41:44 by aben-ham          #+#    #+#             */
-/*   Updated: 2022/02/28 00:08:22 by aben-ham         ###   ########.fr       */
+/*   Updated: 2022/03/01 14:50:53 by aben-ham         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,8 +30,12 @@ int	main(int ac, char **av, char **env)
 	//char		*str;
 	//int			status;
 
-	//char	*str = av[1];
-	char	*str = "|ls $t ";
+	#define D
+	#ifdef D
+	char	*str = av[1];
+	#else
+	char	*str = "o=aasd   ls -l";
+	#endif
 	//process_command(str, env);
 	//process_command(av[1], env);
 	printf("%s\n", str);
