+++ sd
@@ -1,0 +1,14 @@
+Makefile
+_OUT
+a
+commands
+includes
+ls
+main.c
+minishell
+p.py
+parser
+sd
+src
+test
+utils
